Move multiplication to separate modules
